requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-registration==2.4.1
django-tinymce==3.1.0
djangorestframework==3.12.1
Pillow==8.0.1
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
whitenoise==5.2.0
